Timeseries Microservice: Model Registry Integration (#91)
*  Integrated Model Registry with Timeseries Microservice
* Added Model Registry as git submodule
* Updated Makefile for MR build, check and deployment

* Removed depends on from Simulator(opcua, mqtt)
When Timeseries Microservice fails with some issue, simulator stays in always
created state and never goes to deployed state.
Because of this, we have moved the logic to code to check whether
Timeseries Microservice is up

* Moved the build context to parent directory
By changing the build context, this avoids us to use multiple images
to deploy with Helm/docker compose. Since model file is above 3mb
volume mount of the pkl is not supported by helm deployment


---------
Signed-off-by: Vellaisamy, Sathyendran <sathyendran.vellaisamy@intel.com>